+++ .env
@@ -1,6 +1,6 @@
 NEXT_PUBLIC_urlSite=http://localhost:3000/ 
-NEXT_PUBLIC_apiId=328
-NEXT_PUBLIC_apiUrl=https://dev.infoimoveis.com.br/webservice/hotsites.php
+NEXT_PUBLIC_apiBase=https://gateway.marvel.com/v1/public/
+NEXT_PUBLIC_apiUrl=''
 NEXT_PUBLIC_urlImgs=''
 NEXT_PUBLIC_public_key=152afcb329c7159c72b19f0fd2b9d771
 NEXT_PUBLIC_private_key=7ee417ef048c13d2322e4e0dec402f736161464e
