+++ lib/marvelUtils.js
@@ -1,0 +1,6 @@
+//Pela ciencia
+
+export const apiBase    = process.env.NEXT_PUBLIC_apiBase;
+export const urlSite    = process.env.NEXT_PUBLIC_urlSite;
+export const publicKey  = process.env.NEXT_PUBLIC_public_key;
+export const privateKey = process.env.NEXT_PUBLIC_private_key;
