--- lib/tmdb.js
@@ -1,5 +1,0 @@
-//Pela ciencia
-export const apikey ='5a9ac975da01e7ff2c086d9b3076541a';
-export const apiBase = 'https://api.themoviedb.org/3';
-//export const apiLocal = 'https://cinemanext.pages.dev/';
-export const apiLocal = 'http://localhost:3000/';
