+++ package.json
@@ -15,9 +15,9 @@
     "react-dom": "18.2.0"
   },
   "devDependencies": {
-    "@cloudflare/next-on-pages": "^0.0.0-b33639f",
     "@types/node": "18.11.9",
     "@types/react": "18.0.25",
+    "typescript": "^5.1.3",
     "eslint": "8.25.0",
     "eslint-config-next": "12.3.1"
   },
